# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Architecture Knowledge Platform API
  version: 0.2.0
  x-akp-write-idempotency: >
    Mutating POST operations accept Idempotency-Key and MCP write tools require
    it. POST /v1/auth/session is intentionally exempt because it creates a
    Set-Cookie response that cannot be safely replayed.
  description: >
    HTTP contract for the routes registered by the current Fastify runtime.
    All /v1 operations require a scoped bearer API token or an unexpired
    HttpOnly web session and enforce the permission recorded in
    x-akp-permission. Session-authenticated writes also require X-CSRF-Token.
servers:
  - url: http://127.0.0.1:8080
    description: Local development default
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /v1/evals:
    get:
      tags: [Evaluation]
      operationId: listEvaluationRuns
      x-akp-permission: knowledge:read
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/evals/run:
    post:
      tags: [Evaluation]
      operationId: runEvaluation
      x-akp-permission: eval:run
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvaluationTarget" }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/evals/benchmark:
    post:
      tags: [Evaluation]
      operationId: benchmarkRetrieval
      x-akp-permission: eval:run
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvaluationTarget" }
      responses:
        "200": { $ref: "#/components/responses/ObjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
      description: >
        Provisioned out of band with an explicit TokenScopeSet. Token scopes
        are intersected with the user's current memberships on every request.
    cookieAuth:
      type: apiKey
      in: cookie
      name: akp_session
      description: HttpOnly session carrying the effective TokenScopeSet snapshot.
...
  responses:
    ObjectResponse:
      description: Runtime response object
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
...
    Unauthorized:
      description: Missing, expired, revoked or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Permission, space or path scope denied
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code]
      properties:
        code: { type: string }
      additionalProperties: true
...
    EvaluationTarget:
      type: object
      additionalProperties: false
      required: [spaceId, vaultId]
      properties:
        spaceId: { type: string, format: uuid }
        vaultId: { type: string, format: uuid }
        evalPack:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{1,62}$"
          default: generic
...